adding Edge, EdgeWeightedGraph, Kruskals and Prims
Edge – Implements the method and members of this class:
Data is two verts and a weight
Default Constructor that passes in 2 vertices and a weight.
A function that returns one vertex, and another for the other.
compareTo to return the number of differences between one edge to this one
to string that dumps the edge to a string

EdgeWeightedGraph – Implements the graph class, and includes edges
2 constructors that generates space for V number of verts.
Methods to return the number of verts and another for edges
Return adjacency of a vertex
Return a collection of edges
And tostring to dump the graph to strings

Prims- page 622 in the Sedgewick & Wayne text, algorithm 4.7 Eager version
Constructs an mst generated by prim’s algorithm
A help method to keep track of visited nodes
A method to return the edges in a collection
Return the weight of the entire graph

Kruskal’s- page 627 of Sedgewick&Wayne text, algorithm 4.8
Constructs an MST based on kruskal’s algorithm
Method to return the collection of edge
And a method to return the weight