Adding New Files/Resolved Errors
-Added Union Find Class
It contains the code for the implementation as well as a citation for a resource I used/book pages.

-Using java,util.priorityqueue to perform in place of MinPQ. Needed a Comparator class to be implemented to align sorting behavior.

-Implemented IndexedMinPQ
and by implemented, I mean copy pasted of princeton's website, see citation in comments of source file for more

removed errors caused by type mismatches, refencing errors and mis-named call methods.